Got Multiple Elevators Functional
Created a UI button function that allows the player to teleport to each elevator based on a name checking system. The text used for the button will be checked against the dictionary that holds the name for each elevator, if they match the player can teleport to it.

Next is to have the pop-up only show up when players interact with the elevator.